+++ semcore/utils/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.20.1] - 2024-02-05
+
+### Changed
+
+- Deprecated all `@semcore/ui/utils/lib/invalid-state-pattern/InvalidStatePattern`.
+
 ## [4.20.0] - 2024-02-01
 
 ### Added